before refactor to recursive

+++ includes/fillit.h
@@ -33,8 +33,8 @@
 
 typedef struct		s_point
 {
-	int8_t x;
-	int8_t y;
+	int x;
+	int y;
 }					t_point;
 
 typedef	struct		s_map
@@ -44,38 +44,38 @@ typedef	struct		s_map
 
 typedef struct 		s_tris
 {
-	uint8_t			symbol;
-	int8_t			xy[4][2];
+	int			symbol;
+	int			xy[4][2];
 	struct s_tris	*next;
 	struct s_tris	*prev;
 }					t_tris;
 
-t_tris		*ft_read_file_des(int fd);
-t_tris		*ft_lst_new_elem(int8_t xy[4][2], uint8_t symbol);
-t_map		*ft_resize_map(t_map *map);
-t_map		*ft_map(uint16_t size);
-t_map		*ft_solve(t_map *map, t_tris *list);
-t_point		*ft_remove_ttris(t_map **map, char c);
+t_tris	*ft_read_file_des(int fd);
+t_tris	*ft_lst_new_elem(int xy[4][2], int symbol);
+t_map	*ft_resize_map(t_map *map);
+t_map	*ft_map(int size);
+t_map	*ft_solve(t_map *map, t_tris *list);
+t_point	*ft_remove_ttris(t_map **map, char c);
 
-void		ft_lst_push_back(t_tris **head, int8_t xy[4][2], uint8_t symbol);
-void		ft_push_tetris(char *buffer, t_tris **head, uint8_t symbol);
-void    	ft_exit_error();
-void    	ft_display_usage_too_mny_args();
-void	    ft_display_usage_no_file();
-void		ft_exit_invalid_piece();
-void 		ft_print_list(t_tris *list);
-void		ft_lstfree(t_tris *head);
-void		ft_fill_map(char **map, uint16_t sqrt);
-void    	ft_tetris_valid(char *buff);
-void		ft_print_map(char **map);
-void		ft_reformat_coords(int8_t (*xy)[4][2]);
-void		ft_place(t_map *map, t_tris *tris, int8_t x, int8_t y);
+void	ft_lst_push_back(t_tris **head, int xy[4][2], int symbol);
+void	ft_push_tetris(char *buffer, t_tris **head, int symbol);
+void    ft_exit_error();
+void    ft_display_usage_too_mny_args();
+void	ft_display_usage_no_file();
+void	ft_exit_invalid_piece();
+void 	ft_print_list(t_tris *list);
+void	ft_lstfree(t_tris *head);
+void	ft_fill_map(char **map, int sqrt);
+void    ft_tetris_valid(char *buff);
+void	ft_print_map(char **map);
+void	ft_reformat_coords(int (*xy)[4][2]);
+void	ft_place(t_map *map, t_tris *tris, int x, int y);
 
 
-uint16_t	ft_lstlen(t_tris *list);
-uint16_t	ft_size_map(uint16_t size);
-uint8_t 	ft_map_len(char **map);
-int8_t		ft_check(t_map *map, t_tris *tris, int8_t x, int8_t y);
-int			ft_place_tetris(t_map *map, t_tris *tetris_l);
+int		ft_lstlen(t_tris *list);
+int		ft_size_map(int size);
+size_t 	ft_map_len(char **map);
+int		ft_check(t_map *map, t_tris *tris, int x, int y);
+int		ft_place_tetris(t_map *map, t_tris *tetris_l);
 
 #endif
